winegstreamer/video_decoder: Use video_decoder to implement wmv decoder.
(cherry picked from commit e7225aaa94509ba7521993af597dee14510c506e)

CW-Bug-Id: #20833
