+++ dlls/winegstreamer/Makefile.in
@@ -32,7 +32,6 @@ SOURCES = \
 	winegstreamer_classes.idl \
 	wm_reader.c \
 	wma_decoder.c \
-	wmv_decoder.c \
 	media-converter/audioconv.c \
 	media-converter/audioconvbin.c \
 	media-converter/fossilize.c \
